chore(mempool_node): move component execution config
commit-id:4bc56b16
